Sort packages and remove unused vars
