+++ oreClient/package.json
@@ -2,45 +2,45 @@
   "private": true,
   "license": "UNLICENSED",
   "dependencies": {
-    "@fortawesome/free-solid-svg-icons": "5.12.1",
-    "vuex": "3.1.3",
-    "@fortawesome/vue-fontawesome": "0.1.9",
-    "markdown-it-wikilinks": "1.0.1",
-    "vue": "2.6.11",
-    "markdown-it-anchor": "5.2.5",
-    "nprogress": "0.2.0",
-    "@fortawesome/free-regular-svg-icons": "5.12.1",
-    "@fortawesome/fontawesome-svg-core": "1.2.27",
-    "query-string": "6.11.1",
-    "markdown-it-task-lists": "2.1.1",
-    "markdown-it": "10.0.0",
-    "vue-router": "3.1.4",
-    "@fortawesome/free-brands-svg-icons": "5.12.1",
-    "lodash": "4.17.15"
+    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/free-brands-svg-icons": "^5.13.0",
+    "@fortawesome/free-regular-svg-icons": "^5.13.0",
+    "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@fortawesome/vue-fontawesome": "^0.1.9",
+    "lodash": "^4.17.15",
+    "markdown-it": "^10.0.0",
+    "markdown-it-anchor": "^5.2.7",
+    "markdown-it-task-lists": "^2.1.1",
+    "markdown-it-wikilinks": "^1.0.1",
+    "nprogress": "^0.2.0",
+    "query-string": "^6.12.1",
+    "vue": "^2.6.11",
+    "vue-router": "^3.1.6",
+    "vuex": "^3.2.0"
   },
   "devDependencies": {
-    "source-map-loader": "0.2.3",
-    "vue-loader": "15.9.0",
-    "postcss-loader": "3.0.0",
-    "webpack-merge": "4.2.2",
-    "@babel/core": "7.8.7",
-    "node-sass": "4.13.1",
-    "@babel/preset-env": "7.8.7",
-    "copy-webpack-plugin": "5.1.1",
-    "vue-template-compiler": "2.6.11",
-    "webpack": "4.42.0",
-    "concat-with-sourcemaps": "1.0.7",
-    "sass-loader": "8.0.2",
-    "autoprefixer": "9.7.4",
-    "json5-loader": "4.0.0",
-    "webpack-bundle-analyzer": "3.6.1",
-    "webpack-cli": "3.3.2",
-    "vue-style-loader": "4.1.2",
-    "optimize-css-assets-webpack-plugin": "5.0.3",
-    "mini-css-extract-plugin": "0.9.0",
-    "webpack-dev-server": "3.10.3",
-    "babel-loader": "8.0.6",
-    "css-loader": "3.4.2",
-    "terser-webpack-plugin": "2.3.5"
+    "@babel/core": "^7.9.0",
+    "@babel/preset-env": "^7.9.5",
+    "autoprefixer": "^9.7.6",
+    "babel-loader": "^8.1.0",
+    "concat-with-sourcemaps": "^1.1.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "css-loader": "^3.5.2",
+    "json5-loader": "^4.0.0",
+    "mini-css-extract-plugin": "^0.9.0",
+    "node-sass": "^4.13.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "postcss-loader": "^3.0.0",
+    "sass-loader": "^8.0.2",
+    "source-map-loader": "^0.2.4",
+    "terser-webpack-plugin": "^2.3.5",
+    "vue-loader": "^15.9.1",
+    "vue-style-loader": "^4.1.2",
+    "vue-template-compiler": "^2.6.11",
+    "webpack": "^4.42.1",
+    "webpack-bundle-analyzer": "^3.7.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3",
+    "webpack-merge": "^4.2.2"
   }
 }
